Repositories unit tests

+++ Backend/BikesRentalServer.DataAccess/Repositories/BikesRepository.cs
@@ -49,6 +49,9 @@ namespace BikesRentalServer.DataAccess.Repositories
 
         public Bike Remove(Bike entity)
         {
+            if (!_dbContext.Bikes.Contains(entity))
+                return null;
+            
             var bike = _dbContext.Bikes.Remove(entity).Entity;
             _dbContext.SaveChanges();
 
@@ -75,7 +78,7 @@ namespace BikesRentalServer.DataAccess.Repositories
         public Bike Associate(string id, User user)
         {
             var bike = Get(id);
-            if (bike is null)
+            if (bike is null || !_dbContext.Users.Contains(user))
                 return null;
 
             bike.Station = null;
@@ -88,7 +91,7 @@ namespace BikesRentalServer.DataAccess.Repositories
         public Bike Associate(string id, Station station)
         {
             var bike = Get(id);
-            if (bike is null)
+            if (bike is null || !_dbContext.Stations.Contains(station))
                 return null;
 
             bike.User = null;
